+++ packages/create-hydrogen/src/commands/init.ts
@@ -1,0 +1,1 @@
+export {default} from '@shopify/cli-hydrogen/commands/init';